+++ index.html
@@ -452,6 +452,5 @@
       </section>
     </template>
   </body>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </html>
